+++ sources/main.cpp
@@ -3,7 +3,7 @@
 int main() {
     Extract extracteur;
     extracteur.lireFichier("ressources/fourmilieres_texte/fourmiliere_0.txt");    
-    int ant = extracteur.getants();
+    short ant = extracteur.getants();
     const auto& rooms = extracteur.getRooms();
     const auto& connexions = extracteur.getConnexions();
     return 0;
